+++ src/version/version.controller.spec.ts
@@ -1,8 +1,0 @@
-import * as chai from "chai";
-const expect = chai.expect;
-
-describe("VersionController", () => {
-  it("get", () => {
-    expect(true).is.true;
-  });
-});